+++ src/features/AgentBuilder/AgentBuilderConversation.tsx
@@ -33,7 +33,7 @@ const AgentBuilderConversation = memo<AgentBuilderConversationProps>(({ agentId 
         <Flexbox flex={1} style={{ overflow: 'hidden' }}>
           <ChatList welcome={<AgentBuilderWelcome />} />
         </Flexbox>
-        <ChatInput leftActions={actions} rightActions={rightActions} showRuntimeConfig={false} />
+        <ChatInput leftActions={actions} rightActions={rightActions} showControlBar={false} />
       </Flexbox>
     </DragUploadZone>
   );
