+++ src/features/ChatInput/ControlBar/ApprovalMode.tsx
