+++ src/features/ChatInput/ControlBar/HeteroDeviceSwitcher.tsx
