+++ src/features/ChatInput/ControlBar/ModeSelector.tsx
