+++ src/features/ChatInput/ControlBar/WorkingDirectoryPicker.tsx
