+++ src/features/ChatInput/ControlBar/WorkingDirectorySection.tsx
