+++ src/features/ChatInput/ControlBar/dirIcon.tsx
