+++ src/features/ChatInput/ControlBar/useDeviceGitInfo.ts
