+++ src/features/PageEditor/Copilot/Conversation.tsx
@@ -82,7 +82,7 @@ const Conversation = memo(() => {
           leftContent={leftContent}
           sendAreaPrefix={modelSelector}
           sendButtonProps={COMPACT_SEND_BUTTON_PROPS}
-          showRuntimeConfig={false}
+          showControlBar={false}
         />
       </Flexbox>
     </DragUploadZone>
