+++ src/routes/(main)/agent/features/Conversation/WorkingSidebar/index.tsx
@@ -5,7 +5,7 @@ import { lazy, memo } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { DESKTOP_HEADER_ICON_SMALL_SIZE } from '@/const/layoutTokens';
-import { useRepoType } from '@/features/ChatInput/RuntimeConfig/useRepoType';
+import { useRepoType } from '@/features/ChatInput/ControlBar/useRepoType';
 import RightPanel from '@/features/RightPanel';
 import { useAgentStore } from '@/store/agent';
 import {
