+++ src/routes/(main)/home/features/InputArea/index.tsx
@@ -160,7 +160,7 @@ const InputArea = () => {
               dropdownPlacement="bottomLeft"
               inputContainerProps={inputContainerProps}
               placeholder={dailyHint}
-              showRuntimeConfig={false}
+              showControlBar={false}
             />
           </ChatInputProvider>
         </DragUploadZone>
